move classes for streams to a subpackage
